optional add user & move stuff
